refactor: stop the server importing browser feature modules
src/server reached sideways into the SPA feature tree (`@administration`,
`@app`, `@banner`, `@users`), most critically importing the authorization
policy `hasSufficientAdminRole` from `@administration/utils`. That coupled
the Node-typed server project to DOM-typed source, so any DOM-dependent
addition to those modules would break the server at a distance.

Move the shared shapes down into `@virtool/contracts` (administrator
roles + `hasSufficientAdminRole` + `AdministratorPermissions`, the
`Permission` union, banner colors, the SSE domain/message schemas). The
server imports them from the package; each client feature module
re-exports its piece, leaving browser call sites unchanged.

Add a Biome `noRestrictedImports` override on `apps/web/src/server/**`
blocking those four namespaces so the boundary cannot be re-crossed.
