+++ apps/web/package.json
@@ -29,6 +29,7 @@
 		"@tanstack/react-router": "^1.170.17",
 		"@tanstack/react-start": "^1.168.27",
 		"@tanstack/react-virtual": "^3.14.5",
+		"@virtool/contracts": "workspace:*",
 		"@virtool/logger": "workspace:*",
 		"@virtool/sentry": "workspace:*",
 		"bcrypt": "^6.0.0",
