+++ apps/web/src/banner/types.ts
@@ -1,18 +1,7 @@
 import type { UserNested } from "@users/types";
+import { type BannerColor, bannerColors } from "@virtool/contracts";
 
-/** Allowed colors for an instance banner. Kept in lockstep with the
- *  `messagecolor` Postgres enum in `server/db/schema/messages.ts`. */
-export const bannerColors = [
-	"red",
-	"yellow",
-	"blue",
-	"purple",
-	"orange",
-	"grey",
-] as const;
-
-/** One of the allowed banner colors. */
-export type BannerColor = (typeof bannerColors)[number];
+export { type BannerColor, bannerColors };
 
 /** Tailwind background-color class for each banner color. */
 export const bannerColorClasses: Record<BannerColor, string> = {
