+++ biome.json
@@ -71,5 +71,32 @@
 				"organizeImports": "on"
 			}
 		}
-	}
+	},
+	"overrides": [
+		{
+			"includes": ["apps/web/src/server/**"],
+			"linter": {
+				"rules": {
+					"style": {
+						"noRestrictedImports": {
+							"level": "error",
+							"options": {
+								"patterns": [
+									{
+										"group": [
+											"@administration/**",
+											"@app/**",
+											"@banner/**",
+											"@users/**"
+										],
+										"message": "src/server must not import from browser feature modules. The dependency arrow runs client -> @server/*; shared types and helpers belong in @virtool/contracts, which both sides import down into."
+									}
+								]
+							}
+						}
+					}
+				}
+			}
+		}
+	]
 }
