+++ packages/contracts/src/permissions.ts
@@ -1,0 +1,10 @@
+/** A legacy group permission that gates a specific action. */
+export type Permission =
+	| "cancel_job"
+	| "create_ref"
+	| "create_sample"
+	| "modify_hmm"
+	| "modify_subtraction"
+	| "remove_file"
+	| "remove_job"
+	| "upload_file";
